backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.8
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.14.3
black==26.1.0
boto3==1.42.42
botocore==1.42.42
brotli==1.2.0
browser-use==0.11.9
browser-use-sdk==2.0.15
bubus==1.5.6
cdp-use==1.4.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.4
ddgs==9.10.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
ecdsa==0.19.1
email-validator==2.3.0
emergentintegrations==0.1.0
fake-useragent==2.2.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.20.3
flake8==7.3.0
frozenlist==1.8.0
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.62.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
inquirerpy==0.3.4
isort==7.0.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jq==1.11.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.7.8
litellm==1.80.0
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.2
MarkupSafe==3.0.3
mccabe==0.7.0
mcp==1.26.0
mdurl==0.1.2
motor==3.3.1
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.2
oauthlib==3.3.1
ollama==0.6.1
openai==2.21.0
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pathspec==1.0.4
pdfminer.six==20251230
pdfplumber==0.11.9
pfzy==0.3.4
pillow==12.1.0
platformdirs==4.5.1
playwright==1.58.0
pluggy==1.6.0
portalocker==2.10.1
posthog==7.8.6
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.5.0
PyMuPDF==1.27.1
pyotp==2.9.0
pyparsing==3.3.2
pypdf==6.7.0
pypdfium2==5.4.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
screeninfo==0.8.1
shellingham==1.5.4
six==1.17.0
slack_bolt==1.27.0
slack_sdk==3.40.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.3
sse-starlette==3.2.0
starlette==0.52.1
stripe==14.3.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid7==0.1.0
uvicorn==0.40.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
zipfile36==0.1.3
zipp==3.23.0
